% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobR.R
\name{alignmentGroups}
\alias{alignmentGroups}
\title{group fragmented alignments based on proximity to highest score on each subject template}
\usage{
alignmentGroups(BLASTdf, query_length_DNA)
}
\arguments{
\item{BLASTdf}{dataframe of BLAST data from Y1000 project}

\item{query_length_DNA}{DNA length of the query sequence}
}
\description{
group fragmented alignments based on proximity to highest score on each subject template
}
