% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobR.R
\name{angle}
\alias{angle}
\title{Computes the angle between three points}
\usage{
angle(A, B, C, degrees = FALSE)
}
\arguments{
\item{A}{Vector containing the xy-cooydinates of point A}

\item{B}{Vector containing the xy-cooydinates of point B. This point acts as the vertex of angle ABC}

\item{C}{Vector containing the xy-cooydinates of point C}
}
\value{
Angle between the three points in radians
}
\description{
\code{angle} computes the angle between three points
}
