% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-bobhelpR.R
\docType{package}
\name{bobhelpR}
\alias{bobhelpR}
\title{bobhelpR R package}
\description{
bobhelpR R package
}
\author{
Robert J.D. Reid \email{robert.reid381@gmail.edu},

Maintainer: Robert J.D. Reid \email{robert.reid381@gmail.edu}
}
