% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobR.R
\name{colonyCount}
\alias{colonyCount}
\title{Identifies colonies on a scanned plate and returns number of colonies}
\usage{
colonyCount(img, labelText = "", maxArea = 800, minArea = 25,
  eccentricity = 0.6)
}
\arguments{
\item{labelText}{Text to display over image}

\item{maxArea}{Maximum size cutoff to filter out image artefacts}

\item{minArea}{Minimum size cutoff to filter non-colony objects}

\item{eccentricity}{Cutoff for colony objects}

\item{image}{A single plate image}
}
\description{
Identifies colonies on a scanned plate and returns number of colonies
}
