% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobR.R
\name{fineSegment}
\alias{fineSegment}
\title{Return number segments of aligned sequences taking gaps into account}
\usage{
fineSegment(start, query, subject)
}
\arguments{
\item{start}{integer start position of the alignment with respect to the query strand}

\item{query}{string containing query strand sequence from alignment}

\item{subject}{string containing subject strand sequence from alignment}
}
\description{
Return number segments of aligned sequences taking gaps into account
}
