% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobR.R
\name{maxGroupScore}
\alias{maxGroupScore}
\title{find maximum score of alignment groups for each species}
\usage{
maxGroupScore(BLASTdf)
}
\arguments{
\item{BLASTdf}{data frame of Y1000 BLAST server information}
}
\description{
find maximum score of alignment groups for each species
}
