% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobR.R
\name{maxInContig}
\alias{maxInContig}
\title{Find the maximum alignment score on each subject DNA contig contained in the BLAST data}
\usage{
maxInContig(contig_names, align_mid, score)
}
\arguments{
\item{contig_names}{vector of contig names taken from BLAST data}

\item{align_mid}{vector of alignment midpoints from BLAST data}

\item{score}{vector of BLAST scores from BLAST data}
}
\description{
Find the maximum alignment score on each subject DNA contig contained in the BLAST data
}
