% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobR.R
\name{readYKBLAST}
\alias{readYKBLAST}
\title{read BLAST results file from the Y1000 BLAST server}
\usage{
readYKBLAST(file, skip = 3, prefix = "# Fields: ")
}
\arguments{
\item{file}{path to BLAST results text file}

\item{skip}{number of lines to skip to get to header}

\item{prefix}{leading prefix text to remove from header line}
}
\description{
read BLAST results file from the Y1000 BLAST server
}
