% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobR.R
\name{seqPile}
\alias{seqPile}
\title{generate a vector of numbers based on start and end points of sequences}
\usage{
seqPile(start_vector, end_vector)
}
\arguments{
\item{start_vector}{vector of start numbers for generation of sequences}

\item{end_vector}{vector of end numbers for generation of sequences}
}
\description{
generate a vector of numbers based on start and end points of sequences
}
