% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobR.R
\name{sumGroupScores}
\alias{sumGroupScores}
\title{Sum scores of grouped alignments}
\usage{
sumGroupScores(BLASTdf)
}
\arguments{
\item{BLASTdf}{dataframe containing BLAST results from Y1000 server}
}
\description{
Sum scores of grouped alignments
}
