% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobR.R
\name{unGap}
\alias{unGap}
\title{remove sequence insertions from sequence 2 based on gaps in sequence 1}
\usage{
unGap(seq1, seq2)
}
\arguments{
\item{seq1}{string containing a reference sequence}

\item{seq2}{string containing a sequence with insertions to remove}
}
\description{
remove sequence insertions from sequence 2 based on gaps in sequence 1
}
