% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobR.R
\name{gapPositions}
\alias{gapPositions}
\title{find all positions of gaps in a sequence}
\usage{
gapPositions(sequence, gapSize = 1)
}
\arguments{
\item{sequence}{a sequence string with gaps denoted by '-' from a sequence alignment of BLAST search}

\item{gapSize}{integer defining the number of gaps >= gapSize to consider when finding positions}
}
\description{
returns a vector of positions in the sequence string representing sequence position after which a gap of \code{gapSize} or greater occur
numbers returned are for a completely ungapped sequence
}
